Add means for checking that expected objects are contained in the actual array
Quite often it's necessary to check that some items
(objects/strings/arrays etc.) are found in the actual array. With this
patch things become really easy. Previously there was only `contain`
which doesn't support deep equal check for objects.

When using `containAll` the order of the items doesn't matter. This is
great especially when the test can't affect the order that production
code provides (e.g. randomization, shuffle).